+++ tests/includes/rest/class-test-outbox-controller.php
@@ -7,6 +7,7 @@
 
 namespace Activitypub\Tests\Rest;
 
+use Activitypub\Scheduler\Actor;
 use Activitypub\Collection\Outbox;
 use Activitypub\Rest\Outbox_Controller;
 
@@ -78,10 +79,23 @@ class Test_Outbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Tes
 	 * @covers ::validate_user_id
 	 */
 	public function test_validate_user_id() {
+		$actor_mode = \get_option( 'activitypub_actor_mode' );
+		\update_option( 'activitypub_actor_mode', ACTIVITYPUB_ACTOR_AND_BLOG_MODE );
+
 		$controller = new Outbox_Controller();
 		$this->assertTrue( $controller->validate_user_id( 0 ) );
 		$this->assertTrue( $controller->validate_user_id( '1' ) );
 		$this->assertWPError( $controller->validate_user_id( 'user-1' ) );
+
+		\update_option( 'activitypub_actor_mode', ACTIVITYPUB_ACTOR_MODE );
+		$this->assertWPError( $controller->validate_user_id( 0 ) );
+		$this->assertTrue( $controller->validate_user_id( 1 ) );
+
+		\update_option( 'activitypub_actor_mode', ACTIVITYPUB_BLOG_MODE );
+		$this->assertTrue( $controller->validate_user_id( '0' ) );
+		$this->assertWPError( $controller->validate_user_id( 1 ) );
+
+		\update_option( 'activitypub_actor_mode', $actor_mode );
 	}
 
 	/**
@@ -523,6 +537,12 @@ class Test_Outbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Tes
 	 * @covers ::get_items
 	 */
 	public function test_get_items_actor_type_filtering() {
+		\remove_action( 'add_option_activitypub_actor_mode', array( Actor::class, 'blog_user_update' ) );
+		\remove_action( 'update_option_activitypub_actor_mode', array( Actor::class, 'blog_user_update' ) );
+
+		$actor_mode = \get_option( 'activitypub_actor_mode' );
+		\update_option( 'activitypub_actor_mode', ACTIVITYPUB_ACTOR_AND_BLOG_MODE );
+
 		$user_id = self::factory()->user->create( array( 'role' => 'author' ) );
 
 		// Create a post with user actor type.
@@ -530,7 +550,7 @@ class Test_Outbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Tes
 			array(
 				'post_author'  => $user_id,
 				'post_type'    => Outbox::POST_TYPE,
-				'post_status'  => 'draft',
+				'post_status'  => 'pending',
 				'post_title'   => 'https://example.org/activity/1',
 				'post_content' => wp_json_encode(
 					array(
@@ -558,7 +578,7 @@ class Test_Outbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Tes
 			array(
 				'post_author'  => 0,
 				'post_type'    => Outbox::POST_TYPE,
-				'post_status'  => 'draft',
+				'post_status'  => 'pending',
 				'post_title'   => 'https://example.org/activity/2',
 				'post_content' => wp_json_encode(
 					array(
@@ -602,6 +622,10 @@ class Test_Outbox_Controller extends \Activitypub\Tests\Test_REST_Controller_Tes
 		\wp_delete_post( $user_post_id, true );
 		\wp_delete_post( $blog_post_id, true );
 		\wp_delete_user( $user_id );
+
+		\update_option( 'activitypub_actor_mode', $actor_mode );
+		\add_action( 'add_option_activitypub_actor_mode', array( Actor::class, 'blog_user_update' ) );
+		\add_action( 'update_option_activitypub_actor_mode', array( Actor::class, 'blog_user_update' ) );
 	}
 
 	/**