+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Undo API for Outbox items.
 * Metadata to New Follower E-Mail.
 * Allow Activities on URLs instead of requiring Activity-Objects. This is useful especially for sending Announces and Likes.
+* Outbox Activity IDs can now be resolved when the ActivityPub `Accept header is used.
 * Support for incoming `Move` activities and ensure that followed persons are updated accordingly.
 * Labels to add context to visibility settings in the block editor.
 * WP CLI command to reschedule Outbox-Activities.
@@ -32,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Ellipses now display correctly in notification emails for Likes and Reposts.
 * Send Update-Activity when "Actor-Mode" is changed.
 * Added delay to `Announce` Activity from the Blog-Actor, to not have race conditions.
+* `Actor` validation in several REST API endpoints.
 
 ## [5.2.0] - 2025-02-13